Rename function. Move sha to utils class
